Move client deactivation job to clients package
